reducer를 store에서 따로 분리해주었습니다
